search.html 완성
search.html 완성

+++ src/main/java/com/swithus/community/club/repository/search/ClubSearchRepositoryImpl.java
@@ -12,7 +12,7 @@ public class ClubSearchRepositoryImpl extends QuerydslRepositorySupport implemen
     }
 
     @Override
-    public Page<Object[]> clubSearchPage(Pageable pageable, int region, int sports, String keyword) {
+    public Page<Object[]> clubSearchPage(Pageable pageable, Long region, Long sports, String keyword) {
 
         return null;
     }
